requirements.txt
Django==1.6.5
djangorestframework==2.4.0
eav-django==1.4.7
django_autoslug==1.7.2
django_view_shortcuts==1.3.5
